% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FiInstrumentGet}
\alias{DataAccess.FiInstrumentGet}
\title{Retrieve financial instruments data frame from Solvas|Capital, as of the effective date or period}
\usage{
DataAccess.FiInstrumentGet(da_obj, effective_date = NULL,
  effective_period = NULL)
}
\arguments{
\item{da_obj}{- current instance of Solvas|Capital's DataAccess class.}

\item{effective_date}{- effective date to use for schedule data types}

\item{effective_period}{- effective period to use for schedule data types (1=first period)}
}
\value{
dataframe
}
\description{
Takes a DataAccess da_obj and an effective_date or effective_period and returns a data frame
with the instruments.  Schedule properties are coalesced 
to a single value based on the effective_date or effective_period. 
NOTE: EITHER effective_date or effective_period must be populated, the other one must be NULL
}

