% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FsAssumptionsGet}
\alias{DataAccess.FsAssumptionsGet}
\title{Retrieve economic assumptions data frame from Solvas|Capital}
\usage{
DataAccess.FsAssumptionsGet(da_obj, tm_desc = NULL, use_dates = FALSE)
}
\arguments{
\item{da_obj}{- current instance of Solvas|Capital's DataAccess class.}

\item{tm_desc}{- description of transformation (i.e. '(apply to all)'). The first transformation, by sequence 
order, will be used if NULL. If two or more transformations have the same description
 an error message is returned.}

\item{use_dates}{- if TRUE, matrix columns will be dates, else periods. Default to false.}
}
\value{
dataframe
}
\description{
Gets economic schedules from Solvas|Capital transformation data
}

