% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FsAssumptionsScalarGet}
\alias{DataAccess.FsAssumptionsScalarGet}
\title{Retrieve scalar assumption data frame from Solvas|Capital}
\usage{
DataAccess.FsAssumptionsScalarGet(da_obj, tm_desc = NULL)
}
\arguments{
\item{da_obj}{- current instance of Solvas|Capital's DataAccess class.}

\item{tm_desc}{- description of transformation (i.e. '(apply to all)'). The first transformation, by sequence 
order, will be used if NULL. If two or more transformations have the same description
 an error message is returned.}
}
\value{
dataframe (rowname is property_code, columns type_name and property_value )
}
\description{
Gets scalar values from Solvas|Capital transformation data
}
