% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FsScenarioInfoGet}
\alias{DataAccess.FsScenarioInfoGet}
\title{Retrieve general Scenario and Entity information frame from Solvas|Capital, for the scenario ID specified}
\usage{
DataAccess.FsScenarioInfoGet(da_obj)
}
\arguments{
\item{da_obj}{- current instance of Solvas|Capital's DataAccess class.}
}
\value{
dataframe
}
\description{
Takes a DataAccess da_obj and returns a dataframe with general information about the scenario including:

entity_start_date - Start date of the entity \cr
entity_end_date - End date of the entity \cr
effective_start_date - Effective date of the scenario \cr
effective_end_date - End Date of the last reporting period defined for the entity \cr
scenario_description - Name of the Scenario \cr
functional_currency - Currency specified for the Entity \cr
coa_description - Name of the Chart of Accounts associated with the Entity \cr
reporting_period_length - Length of reporting periods for the entity (Month, Year) \cr
first_reporting_period_end_date - End date of first reporting period for the entity \cr
first_reporting_year_end_date - End date of first reporting year for the entity \cr

Example:
  scenario_info <- DataAccess.FsScenarioInfoGet(sc_da) \cr
  Currency <- scenario_info$functional_currency \cr
}

