% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.PackageVersionCompatible}
\alias{DataAccess.PackageVersionCompatible}
\title{Check version compatibility.}
\usage{
DataAccess.PackageVersionCompatible(da_obj)
}
\arguments{
\item{da_obj}{- current instance of Solvas|Capital's DataAccess class.}
}
\value{
"success" confirms the version is compatible with the Solvas|Capital database server, otherwise an error message is returned.
}
\description{
Takes a DataAccess da_obj and returns "success" if the connection is working, or an error if the version 
is incompatible.
}

