% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSAssumptionsScalarGet}
\alias{SPFSAssumptionsScalarGet}
\title{Internal Solvas|Capital Function 
gets Solvas|Capital assumptions}
\usage{
SPFSAssumptionsScalarGet(connection_string, fs_id, tm_desc)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{tm_desc}{- description/name of transformation (i.e. '(apply to all)').  the first}
}
\value{
data frame containing the economic factors
}
\description{
Internal Solvas|Capital Function 
gets Solvas|Capital assumptions
}
\keyword{internal}
