% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSCreateInstruments}
\alias{SPFSCreateInstruments}
\title{Internal Solvas|Capital Function 
Creates instruments/properites for the passed fs_id}
\usage{
SPFSCreateInstruments(connection_string, fs_id)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{tm_desc}{- description/name of transformation (i.e. '(apply to all)').  the first}
}
\value{
data frame containing the economic factors
}
\description{
Internal Solvas|Capital Function 
Creates instruments/properites for the passed fs_id
}
\keyword{internal}

