% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSScenarioInfoGet}
\alias{SPFSScenarioInfoGet}
\title{Internal Solvas|Capital Function 
Get scenario/entity information}
\usage{
SPFSScenarioInfoGet(connection_string, fs_id)
}
\arguments{
\item{connection_string}{The string representing the connection string...}

\item{fs_id}{The scenario ID}
}
\value{
data frame containing the general scenario information
}
\description{
Internal Solvas|Capital Function 
Get scenario/entity information
}
\keyword{internal}

