% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess}
\alias{DataAccess}
\title{Access and interact with Solvas|Capital database (da_obj).}
\usage{
DataAccess(connection_string_param = "", fs_id_param = NULL,
  event_id_param = NULL)
}
\arguments{
\item{connection_string_param}{-  SQL Server connection string}

\item{fs_id_param}{- Solvas|Capital scenario ID used in the run}

\item{event_id_param}{- Solvas|Capital event ID of the scenario run (passed from the server or NULL for local development).}
}
\description{
DataAccess is a required object which is always the first parameter passed to any of the other functions.
}
\examples{
\dontrun{
connection_string = "Driver={Sql Server};server=(local);trusted_connection=True;
database=Internal_Capital_DEV;"
sc_da <- DataAccess(connection_string_param=connection_string, fs_id_param=1)
DataAccess.FiInstrumentGet(sc_da, NULL, 1)
}
}
